Extract GameMenu.java from TripleAMenu and BasicGameMenuBar
