+++ src/games/strategy/triplea/ui/menubar/TripleAMenu.java
@@ -95,34 +95,9 @@ public class TripleAMenu extends BasicGameMenuBar {
   @Override
   protected void createGameSpecificMenus(final JMenuBar menuBar) {
     new ViewMenu(menuBar, frame, getData());
-    menuBar.add(createGameMenu());
+    new GameMenu(menuBar, frame, getData());
     createExportMenu(menuBar);
   }
-
-  private JMenu createGameMenu() {
-    final JMenu menuGame = SwingComponents.newJMenu("Game", SwingComponents.KeyboardCode.G);
-    addEditMode(menuGame);
-    menuGame.add(frame.getShowGameAction()).setMnemonic(KeyEvent.VK_G);
-    menuGame.add(frame.getShowHistoryAction()).setMnemonic(KeyEvent.VK_H);
-    menuGame.add(frame.getShowMapOnlyAction()).setMnemonic(KeyEvent.VK_M);
-    addShowVerifiedDice(menuGame);
-    SoundOptions.addGlobalSoundSwitchMenu(menuGame);
-    SoundOptions.addToMenu(menuGame, SoundPath.SoundType.TRIPLEA);
-    menuGame.addSeparator();
-    addGameOptionsMenu(menuGame);
-    addPoliticsMenu(menuGame);
-    addNotificationSettings(menuGame);
-    addFocusOnCasualties(menuGame);
-    addConfirmBattlePhases(menuGame);
-    addShowEnemyCasualties(menuGame);
-    addShowAIBattles(menuGame);
-    addAISleepDuration(menuGame);
-    addShowDiceStats(menuGame);
-    addRollDice(menuGame);
-    addBattleCalculatorMenu(menuGame);
-    return menuGame;
-  }
-
 
   private void createExportMenu(final JMenuBar menuBar) {
     final JMenu menuGame = new JMenu("Export");
@@ -136,220 +111,12 @@ public class TripleAMenu extends BasicGameMenuBar {
     addSaveScreenshot(menuGame);
   }
 
-  private void addEditMode(final JMenu parentMenu) {
-    final JCheckBoxMenuItem editMode = new JCheckBoxMenuItem("Enable Edit Mode");
-    editMode.setModel(frame.getEditModeButtonModel());
-    parentMenu.add(editMode).setMnemonic(KeyEvent.VK_E);
-  }
-
-
-
-  private void addShowVerifiedDice(final JMenu parentMenu) {
-    final Action showVerifiedDice = SwingAction.of("Show Verified Dice..",
-        e -> new VerifiedRandomNumbersDialog(frame.getRootPane()).setVisible(true));
-    if (getGame() instanceof ClientGame) {
-      parentMenu.add(showVerifiedDice).setMnemonic(KeyEvent.VK_V);
-    }
-  }
 
   private void addSaveScreenshot(final JMenu parentMenu) {
     parentMenu.add(frame.getSaveScreenshotAction()).setMnemonic(KeyEvent.VK_E);
   }
 
 
-  private static void addShowEnemyCasualties(final JMenu parentMenu) {
-    final JCheckBoxMenuItem showEnemyCasualties = new JCheckBoxMenuItem("Confirm Enemy Casualties");
-    showEnemyCasualties.setMnemonic(KeyEvent.VK_E);
-    showEnemyCasualties.setSelected(BattleDisplay.getShowEnemyCasualtyNotification());
-    showEnemyCasualties.addActionListener(
-        SwingAction.of(e -> BattleDisplay.setShowEnemyCasualtyNotification(showEnemyCasualties.isSelected())));
-    parentMenu.add(showEnemyCasualties);
-  }
-
-  private static void addFocusOnCasualties(final JMenu parentMenu) {
-    final JCheckBoxMenuItem focusOnCasualties = new JCheckBoxMenuItem("Focus On Own Casualties");
-    focusOnCasualties.setSelected(BattleDisplay.getFocusOnOwnCasualtiesNotification());
-    focusOnCasualties.addActionListener(
-        SwingAction.of(e -> BattleDisplay.setFocusOnOwnCasualtiesNotification(focusOnCasualties.isSelected())));
-    parentMenu.add(focusOnCasualties);
-  }
-
-  private static void addConfirmBattlePhases(final JMenu parentMenu) {
-    final JCheckBoxMenuItem confirmPhases = new JCheckBoxMenuItem("Confirm Defensive Rolls");
-    confirmPhases.setSelected(BattleDisplay.getConfirmDefensiveRolls());
-    confirmPhases.addActionListener(
-        SwingAction.of(e -> BattleDisplay.setConfirmDefensiveRolls(confirmPhases.isSelected())));
-    parentMenu.add(confirmPhases);
-  }
-
-
-  /**
-   * Add a Politics Panel button to the game menu, this panel will show the
-   * current political landscape as a reference, no actions on this panel.
-   *
-   * @param menuGame
-   */
-  private void addPoliticsMenu(final JMenu menuGame) {
-    final AbstractAction politicsAction = SwingAction.of("Show Politics Panel", e -> {
-      final PoliticalStateOverview ui = new PoliticalStateOverview(getData(), getUIContext(), false);
-      final JScrollPane scroll = new JScrollPane(ui);
-      scroll.setBorder(BorderFactory.createEmptyBorder());
-      final Dimension screenResolution = Toolkit.getDefaultToolkit().getScreenSize();
-      // not only do we have a start bar, but we also have the message dialog to account for
-      final int availHeight = screenResolution.height - 120;
-      // just the scroll bars plus the window sides
-      final int availWidth = screenResolution.width - 40;
-
-      scroll.setPreferredSize(
-          new Dimension((scroll.getPreferredSize().width > availWidth ? availWidth : scroll.getPreferredSize().width),
-              (scroll.getPreferredSize().height > availHeight ? availHeight : scroll.getPreferredSize().height)));
-
-      JOptionPane.showMessageDialog(frame, scroll, "Politics Panel", JOptionPane.PLAIN_MESSAGE);
-
-    });
-    menuGame.add(politicsAction).setMnemonic(KeyEvent.VK_P);
-  }
-
-  private void addNotificationSettings(final JMenu parentMenu) {
-    final JMenu notificationMenu = new JMenu();
-    notificationMenu.setMnemonic(KeyEvent.VK_U);
-    notificationMenu.setText("User Notifications...");
-    final JCheckBoxMenuItem showEndOfTurnReport = new JCheckBoxMenuItem("Show End of Turn Report");
-    showEndOfTurnReport.setMnemonic(KeyEvent.VK_R);
-    final JCheckBoxMenuItem showTriggeredNotifications = new JCheckBoxMenuItem("Show Triggered Notifications");
-    showTriggeredNotifications.setMnemonic(KeyEvent.VK_T);
-    final JCheckBoxMenuItem showTriggerChanceSuccessful =
-        new JCheckBoxMenuItem("Show Trigger/Condition Chance Roll Successful");
-    showTriggerChanceSuccessful.setMnemonic(KeyEvent.VK_S);
-    final JCheckBoxMenuItem showTriggerChanceFailure =
-        new JCheckBoxMenuItem("Show Trigger/Condition Chance Roll Failure");
-    showTriggerChanceFailure.setMnemonic(KeyEvent.VK_F);
-    notificationMenu.addMenuListener(new MenuListener() {
-      @Override
-      public void menuSelected(final MenuEvent e) {
-        showEndOfTurnReport.setSelected(getUIContext().getShowEndOfTurnReport());
-        showTriggeredNotifications.setSelected(getUIContext().getShowTriggeredNotifications());
-        showTriggerChanceSuccessful.setSelected(getUIContext().getShowTriggerChanceSuccessful());
-        showTriggerChanceFailure.setSelected(getUIContext().getShowTriggerChanceFailure());
-      }
-
-      @Override
-      public void menuDeselected(final MenuEvent e) {}
-
-      @Override
-      public void menuCanceled(final MenuEvent e) {}
-    });
-    showEndOfTurnReport.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        getUIContext().setShowEndOfTurnReport(showEndOfTurnReport.isSelected());
-      }
-    });
-    showTriggeredNotifications.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        getUIContext().setShowTriggeredNotifications(showTriggeredNotifications.isSelected());
-      }
-    });
-    showTriggerChanceSuccessful.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        getUIContext().setShowTriggerChanceSuccessful(showTriggerChanceSuccessful.isSelected());
-      }
-    });
-    showTriggerChanceFailure.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        getUIContext().setShowTriggerChanceFailure(showTriggerChanceFailure.isSelected());
-      }
-    });
-    notificationMenu.add(showEndOfTurnReport);
-    notificationMenu.add(showTriggeredNotifications);
-    notificationMenu.add(showTriggerChanceSuccessful);
-    notificationMenu.add(showTriggerChanceFailure);
-    parentMenu.add(notificationMenu);
-  }
-
-  private void addShowAIBattles(final JMenu parentMenu) {
-    final JCheckBoxMenuItem showAIBattlesBox = new JCheckBoxMenuItem("Show Battles Between AIs");
-    showAIBattlesBox.setMnemonic(KeyEvent.VK_A);
-    showAIBattlesBox.setSelected(getUIContext().getShowBattlesBetweenAIs());
-    showAIBattlesBox.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        getUIContext().setShowBattlesBetweenAIs(showAIBattlesBox.isSelected());
-      }
-    });
-    parentMenu.add(showAIBattlesBox);
-  }
-
-  private void addShowDiceStats(final JMenu parentMenu) {
-    final Action showDiceStats = SwingAction.of("Show Dice Stats...", e -> {
-      final IRandomStats randomStats =
-          (IRandomStats) getGame().getRemoteMessenger().getRemote(IRandomStats.RANDOM_STATS_REMOTE_NAME);
-      final RandomStatsDetails stats = randomStats.getRandomStats(getData().getDiceSides());
-      JOptionPane.showMessageDialog(frame, new JScrollPane(stats.getAllStats()), "Random Stats",
-          JOptionPane.INFORMATION_MESSAGE);
-    });
-    parentMenu.add(showDiceStats).setMnemonic(KeyEvent.VK_D);
-  }
-
-  private void addRollDice(final JMenu parentMenu) {
-    final JMenuItem RollDiceBox = new JMenuItem("Roll Dice...");
-    RollDiceBox.setMnemonic(KeyEvent.VK_R);
-    RollDiceBox.addActionListener(new ActionListener() {
-      @Override
-      public void actionPerformed(final ActionEvent e) {
-        final IntTextField numberOfText = new IntTextField(0, 100);
-        final IntTextField diceSidesText = new IntTextField(1, 200);
-        numberOfText.setText(String.valueOf(0));
-        diceSidesText.setText(String.valueOf(getGame().getData().getDiceSides()));
-        final JPanel panel = new JPanel();
-        panel.setLayout(new GridBagLayout());
-        panel.add(new JLabel("Number of Dice to Roll: "), new GridBagConstraints(0, 0, 1, 1, 0, 0,
-            GridBagConstraints.WEST, GridBagConstraints.BOTH, new Insets(0, 0, 0, 20), 0, 0));
-        panel.add(new JLabel("Sides on the Dice: "), new GridBagConstraints(2, 0, 1, 1, 0, 0, GridBagConstraints.WEST,
-            GridBagConstraints.BOTH, new Insets(0, 20, 0, 10), 0, 0));
-        panel.add(numberOfText, new GridBagConstraints(0, 1, 1, 1, 0, 0, GridBagConstraints.WEST,
-            GridBagConstraints.BOTH, new Insets(0, 0, 0, 20), 0, 0));
-        panel.add(diceSidesText, new GridBagConstraints(2, 1, 1, 1, 0, 0, GridBagConstraints.WEST,
-            GridBagConstraints.BOTH, new Insets(0, 20, 0, 10), 0, 0));
-        JOptionPane.showOptionDialog(JOptionPane.getFrameForComponent(TripleAMenu.this), panel, "Roll Dice",
-            JOptionPane.OK_OPTION, JOptionPane.INFORMATION_MESSAGE, null, new String[] {"OK"}, "OK");
-        try {
-          final int numberOfDice = Integer.parseInt(numberOfText.getText());
-          if (numberOfDice > 0) {
-            final int diceSides = Integer.parseInt(diceSidesText.getText());
-            final int[] dice =
-                getGame().getRandomSource().getRandom(diceSides, numberOfDice, "Rolling Dice, no effect on game.");
-            final JPanel panelDice = new JPanel();
-            final BoxLayout layout = new BoxLayout(panelDice, BoxLayout.Y_AXIS);
-            panelDice.setLayout(layout);
-            final JLabel label = new JLabel("Rolls (no effect on game): ");
-            panelDice.add(label);
-            String diceString = "";
-            for (int i = 0; i < dice.length; i++) {
-              diceString += String.valueOf(dice[i] + 1) + ((i == dice.length - 1) ? "" : ", ");
-            }
-            final JTextField diceList = new JTextField(diceString);
-            diceList.setEditable(false);
-            panelDice.add(diceList);
-            JOptionPane.showMessageDialog(frame, panelDice, "Dice Rolled", JOptionPane.INFORMATION_MESSAGE);
-          }
-        } catch (final Exception ex) {
-        }
-      }
-    });
-    parentMenu.add(RollDiceBox);
-  }
-
-  private void addBattleCalculatorMenu(final JMenu menuGame) {
-    final Action showBattleMenu = SwingAction.of("Battle Calculator...", e -> OddsCalculatorDialog.show(frame, null));
-    final JMenuItem showBattleMenuItem = menuGame.add(showBattleMenu);
-    showBattleMenuItem.setMnemonic(KeyEvent.VK_B);
-    showBattleMenuItem.setAccelerator(
-        KeyStroke.getKeyStroke(KeyEvent.VK_B, java.awt.Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()));
-  }
 
   private void addExportStatsFull(final JMenu parentMenu) {
     final Action showDiceStats = SwingAction.of("Export Full Game Stats...", e -> createAndSaveStats(true));
